refactor: connetion_type.goをenumに移動
